Move generated bebop.ts to ./src
In preparation for adding more broader bebop functionality
